+++ tsm_exporter_test.go
@@ -65,13 +65,13 @@ FOO,Future
 BAR,Not Started
 BAR,Not Started
 `
-	mockReplicationViewStdout = `
-COMPLETE,2020-03-23 06:06:45.000000,/TEST2CONF,TEST2DB2,2020-03-23 00:45:29.000000,167543418,2
-COMPLETE,2020-03-23 06:06:45.000000,/TEST4,TEST2DB2,2020-03-23 00:45:29.000000,1052637876956,2
-COMPLETE,2020-03-23 00:06:07.000000,/srv,TEST2DB.DOMAIN,2020-03-23 00:05:24.000000,245650752,10
-COMPLETE,2020-03-22 06:02:38.000000,/TEST2CONF,TEST2DB2,2020-03-22 00:45:29.000000,167543418,2
-COMPLETE,2020-03-22 06:02:38.000000,/TEST4,TEST2DB2,2020-03-22 00:45:29.000000,1052637876316,2
-COMPLETE,2020-03-22 00:05:57.000000,/srv,TEST2DB.DOMAIN,2020-03-22 00:05:23.000000,234680204,12
+	mockReplicationViewCompletedStdout = `
+TEST2DB2,/TEST2CONF,2020-03-23 00:45:29.000000,2020-03-23 06:06:45.000000,2,167543418
+TEST2DB2,/TEST4,2020-03-23 00:45:29.000000,2020-03-23 06:06:45.000000,2,1052637876956
+`
+	mockReplicationViewNotCompletedStdout = `
+FOO,/BAR
+BAR,/BAZ
 `
 )
 
@@ -121,8 +121,11 @@ func TestMetricsHandler(t *testing.T) {
 	collector.DsmadmcEventsNotCompletedExec = func(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
 		return mockEventNotCompletedStdout, nil
 	}
-	collector.DsmadmcReplicationViewsExec = func(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
-		return mockReplicationViewStdout, nil
+	collector.DsmadmcReplicationViewsCompletedExec = func(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
+		return mockReplicationViewCompletedStdout, nil
+	}
+	collector.DsmadmcReplicationViewsNotCompletedExec = func(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
+		return mockReplicationViewNotCompletedStdout, nil
 	}
 	body, err := queryExporter("target=test1", http.StatusOK)
 	if err != nil {
@@ -158,8 +161,11 @@ func TestMetricsHandlerCollectorsDefined(t *testing.T) {
 	collector.DsmadmcEventsNotCompletedExec = func(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
 		return mockEventNotCompletedStdout, nil
 	}
-	collector.DsmadmcReplicationViewsExec = func(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
-		return mockReplicationViewStdout, nil
+	collector.DsmadmcReplicationViewsCompletedExec = func(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
+		return mockReplicationViewCompletedStdout, nil
+	}
+	collector.DsmadmcReplicationViewsNotCompletedExec = func(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
+		return mockReplicationViewNotCompletedStdout, nil
 	}
 	body, err := queryExporter("target=test2", http.StatusOK)
 	if err != nil {