+++ collector/events.go
@@ -14,7 +14,6 @@
 package collector
 
 import (
-	"context"
 	"fmt"
 	"strings"
 	"sync"
@@ -69,12 +68,9 @@ func (c *EventsCollector) Describe(ch chan<- *prometheus.Desc) {
 func (c *EventsCollector) Collect(ch chan<- prometheus.Metric) {
 	level.Debug(c.logger).Log("msg", "Collecting metrics")
 	collectTime := time.Now()
-	timeout := 0
 	errorMetric := 0
 	metrics, err := c.collect()
-	if err == context.DeadlineExceeded {
-		timeout = 1
-	} else if err != nil {
+	if err != nil {
 		level.Error(c.logger).Log("msg", err)
 		errorMetric = 1
 	}
@@ -85,24 +81,21 @@ func (c *EventsCollector) Collect(ch chan<- prometheus.Metric) {
 	}
 
 	ch <- prometheus.MustNewConstMetric(collectError, prometheus.GaugeValue, float64(errorMetric), "events")
-	ch <- prometheus.MustNewConstMetric(collecTimeout, prometheus.GaugeValue, float64(timeout), "events")
 	ch <- prometheus.MustNewConstMetric(collectDuration, prometheus.GaugeValue, time.Since(collectTime).Seconds(), "events")
 }
 
 func (c *EventsCollector) collect() (map[string]EventMetric, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), time.Duration(*eventsTimeout)*time.Second)
-	defer cancel()
 	var completedOut, notCompletedOut string
 	var completedErr, notCompletedErr error
 	wg := &sync.WaitGroup{}
 	wg.Add(2)
 	go func() {
 		defer wg.Done()
-		completedOut, completedErr = DsmadmcEventsCompletedExec(c.target, ctx, c.logger)
+		completedOut, completedErr = DsmadmcEventsCompletedExec(c.target, c.logger)
 	}()
 	go func() {
 		defer wg.Done()
-		notCompletedOut, notCompletedErr = DsmadmcEventsNotCompletedExec(c.target, ctx, c.logger)
+		notCompletedOut, notCompletedErr = DsmadmcEventsNotCompletedExec(c.target, c.logger)
 	}()
 	wg.Wait()
 	if completedErr != nil {
@@ -124,8 +117,8 @@ func buildEventsCompletedQuery(target *config.Target) string {
 	return query
 }
 
-func dsmadmcEventsCompleted(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
-	out, err := dsmadmcQuery(target, buildEventsCompletedQuery(target), ctx, logger)
+func dsmadmcEventsCompleted(target *config.Target, logger log.Logger) (string, error) {
+	out, err := dsmadmcQuery(target, buildEventsCompletedQuery(target), *eventsTimeout, logger)
 	return out, err
 }
 
@@ -141,8 +134,8 @@ func buildEventsNotCompletedQuery(target *config.Target) string {
 	return query
 }
 
-func dsmadmcEventsNotCompleted(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
-	out, err := dsmadmcQuery(target, buildEventsNotCompletedQuery(target), ctx, logger)
+func dsmadmcEventsNotCompleted(target *config.Target, logger log.Logger) (string, error) {
+	out, err := dsmadmcQuery(target, buildEventsNotCompletedQuery(target), *eventsTimeout, logger)
 	return out, err
 }