+++ collector/occupancy.go
@@ -14,7 +14,6 @@
 package collector
 
 import (
-	"context"
 	"fmt"
 	"reflect"
 	"sort"
@@ -86,12 +85,9 @@ func (c *OccupancysCollector) Describe(ch chan<- *prometheus.Desc) {
 func (c *OccupancysCollector) Collect(ch chan<- prometheus.Metric) {
 	level.Debug(c.logger).Log("msg", "Collecting metrics")
 	collectTime := time.Now()
-	timeout := 0
 	errorMetric := 0
 	metrics, err := c.collect()
-	if err == context.DeadlineExceeded {
-		timeout = 1
-	} else if err != nil {
+	if err != nil {
 		level.Error(c.logger).Log("msg", err)
 		errorMetric = 1
 	}
@@ -103,14 +99,11 @@ func (c *OccupancysCollector) Collect(ch chan<- prometheus.Metric) {
 	}
 
 	ch <- prometheus.MustNewConstMetric(collectError, prometheus.GaugeValue, float64(errorMetric), "occupancy")
-	ch <- prometheus.MustNewConstMetric(collecTimeout, prometheus.GaugeValue, float64(timeout), "occupancy")
 	ch <- prometheus.MustNewConstMetric(collectDuration, prometheus.GaugeValue, time.Since(collectTime).Seconds(), "occupancy")
 }
 
 func (c *OccupancysCollector) collect() ([]OccupancyMetric, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), time.Duration(*occupancyTimeout)*time.Second)
-	defer cancel()
-	out, err := DsmadmcOccupancysExec(c.target, ctx, c.logger)
+	out, err := DsmadmcOccupancysExec(c.target, c.logger)
 	if err != nil {
 		return nil, err
 	}
@@ -118,7 +111,7 @@ func (c *OccupancysCollector) collect() ([]OccupancyMetric, error) {
 	return metrics, nil
 }
 
-func dsmadmcOccupancys(target *config.Target, ctx context.Context, logger log.Logger) (string, error) {
+func dsmadmcOccupancys(target *config.Target, logger log.Logger) (string, error) {
 	fields := getOccupancyFields()
 	var queryFields []string
 	var groupFields []string
@@ -133,7 +126,7 @@ func dsmadmcOccupancys(target *config.Target, ctx context.Context, logger log.Lo
 		queryFields = append(queryFields, field)
 	}
 	query := fmt.Sprintf("SELECT %s FROM occupancy GROUP BY %s", strings.Join(queryFields, ","), strings.Join(groupFields, ","))
-	out, err := dsmadmcQuery(target, query, ctx, logger)
+	out, err := dsmadmcQuery(target, query, *occupancyTimeout, logger)
 	return out, err
 }