+++ collector/volumes.go
@@ -60,7 +60,7 @@ func NewVolumesExporter(target *config.Target, logger log.Logger) Collector {
 		readonly: prometheus.NewDesc(prometheus.BuildFQName(namespace, "volumes", "readonly"),
 			"Number of readonly volumes", nil, nil),
 		utilized: prometheus.NewDesc(prometheus.BuildFQName(namespace, "volume", "utilized_percent"),
-			"Volume percent utilized", []string{"volume", "classname"}, nil),
+			"Volume percent utilized (ratio of 0.0-1.0)", []string{"volume", "classname"}, nil),
 		capacity: prometheus.NewDesc(prometheus.BuildFQName(namespace, "volume", "estimated_capacity_bytes"),
 			"Volume estimated capacity", []string{"volume", "classname"}, nil),
 		target: target,
@@ -150,7 +150,7 @@ func volumesParse(out string, logger log.Logger) []VolumeMetric {
 			level.Error(logger).Log("msg", "Error parsing pct_utilized value", "value", items[2], "err", err)
 			continue
 		}
-		metric.utilized = utilized
+		metric.utilized = utilized / 100
 		metrics = append(metrics, metric)
 	}
 	return metrics